requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-dialogflow==2.28.0
protobuf>=4.21.6

